requirements.txt
chromadb==1.0.20
sentence-transformers==5.1.0
numpy==2.0.2
langchain == 0.3.27
pypdf2 == 3.0.1
langgraph == 0.6.6